[NEW] Option to switch EEG time unit
The x-axis tick labels of the panel showing EEG traces can now be set to either show the time in seconds, minutes, or hours. This option is available in the configuration window and stores a new item "EEG_panel_time_unit" in the configuration file. A new script takes care that old configuration files are compatible with this new option. Scoring Hero Version 0.0.7.